refactored enums
