+++ hms-app/src/components/card-details/HackathonDetails.tsx
@@ -4,7 +4,7 @@ import {
   getHackathonDetails,
 } from '../../actions/HackathonActions'
 import IdeaCardList from '../lists/IdeaCardList'
-import { Hackathon, Idea } from '../../common/types'
+import { Hackathon, HackathonDetailsType, Idea } from '../../common/types'
 import { getIdeaDetails } from '../../actions/IdeaActions'
 import {
   Accordion,
@@ -48,11 +48,6 @@ const useStyles = createStyles((theme) => ({
 type IProps = {
   hackathonId: string
   type: HackathonDetailsType
-}
-
-export enum HackathonDetailsType {
-  Header = 'HEADER',
-  FullInfo = 'FULLINFO',
 }
 
 export default function HackathonDetails(props: IProps) {
